moved state and communication to utilities
