+++ device/peripherals/classes/atlas/driver.py
@@ -6,10 +6,10 @@ from typing import Optional, Tuple, Dict, NamedTuple
 
 # Import device utilities
 from device.utilities.logger import Logger
-from device.communication.i2c.main import I2C
-from device.communication.i2c.exceptions import I2CError
-from device.communication.i2c.mux_simulator import MuxSimulator
-from device.communication.i2c.peripheral_simulator import PeripheralSimulator
+from device.utilities.communication.i2c.main import I2C
+from device.utilities.communication.i2c.exceptions import I2CError
+from device.utilities.communication.i2c.mux_simulator import MuxSimulator
+from device.utilities.communication.i2c.peripheral_simulator import PeripheralSimulator
 
 # Import manager elements
 from device.peripherals.classes.atlas import exceptions
