+++ device/peripherals/common/dac5578/simulator.py
@@ -5,7 +5,7 @@ from typing import Any, Dict
 from device.utilities.bitwise import byte_str
 
 # Import simulator elements
-from device.communication.i2c.peripheral_simulator import PeripheralSimulator
+from device.utilities.communication.i2c.peripheral_simulator import PeripheralSimulator
 
 
 class DAC5578Simulator(PeripheralSimulator):  # type: ignore
