+++ device/peripherals/modules/atlas_ec/driver.py
@@ -6,9 +6,9 @@ from typing import Optional, Tuple, NamedTuple
 
 # Import device utilities
 from device.utilities import maths
-from device.communication.i2c.main import I2C
-from device.communication.i2c.exceptions import I2CError
-from device.communication.i2c.mux_simulator import MuxSimulator
+from device.utilities.communication.i2c.main import I2C
+from device.utilities.communication.i2c.exceptions import I2CError
+from device.utilities.communication.i2c.mux_simulator import MuxSimulator
 
 # Import module elements
 from device.peripherals.classes.atlas import driver
