+++ device/peripherals/modules/atlas_ec/tests/test_driver.py
@@ -5,7 +5,7 @@ import os, sys, pytest, threading
 sys.path.append(os.environ["OPENAG_BRAIN_ROOT"])
 
 # Import mux simulator
-from device.communication.i2c.mux_simulator import MuxSimulator
+from device.utilities.communication.i2c.mux_simulator import MuxSimulator
 
 # Import peripheral driver
 from device.peripherals.modules.atlas_ec.driver import AtlasECDriver
