+++ device/peripherals/modules/led_dac5578/tests/test_manager.py
@@ -8,8 +8,8 @@ os.chdir(ROOT_DIR)
 
 # Import device utilities
 from device.utilities import accessors
-from device.communication.i2c.mux_simulator import MuxSimulator
-from device.state.main import State
+from device.utilities.communication.i2c.mux_simulator import MuxSimulator
+from device.utilities.state.main import State
 
 # Import peripheral manager
 from device.peripherals.modules.led_dac5578.manager import LEDDAC5578Manager
