+++ device/peripherals/modules/t6713/driver.py
@@ -6,9 +6,9 @@ from typing import NamedTuple, Optional, Tuple
 
 # Import device utilities
 from device.utilities import logger, bitwise
-from device.communication.i2c.main import I2C
-from device.communication.i2c.mux_simulator import MuxSimulator
-from device.communication.i2c.exceptions import I2CError
+from device.utilities.communication.i2c.main import I2C
+from device.utilities.communication.i2c.mux_simulator import MuxSimulator
+from device.utilities.communication.i2c.exceptions import I2CError
 
 # Import driver elements
 from device.peripherals.modules.t6713 import exceptions, simulator
