+++ device/utilities/communication/i2c/exceptions.py
